// src/server/api/models/accounting/models/FixedAssetGroup.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AmortizationFrequency } from './AmortizationFrequency';
import type { LocalDateTime } from './LocalDateTime';
import type { UUID } from './UUID';

export type FixedAssetGroup = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * A unique transaction code generated by the system
     */
    code: string;
    /**
     * A reference used to post the journals
     */
    reference: string;
    /**
     * The amortization frequency
     */
    frequency: AmortizationFrequency;
    /**
     * The months of useful life
     */
    assetLifespan: number;
    /**
     * Notes related to the group
     */
    notes?: string | null;
    /**
     * The transaction currency.
     */
    currency: string;
    /**
     * The residual value of the asset
     */
    residualValue: number;
    /**
     * The record label
     */
    label: string;
    /**
     * The asset account used for processing the acquisition
     */
    assetAccountId: UUID;
    /**
     * The transaction source account
     */
    transactionSourceAccountId: UUID;
    /**
     * The accumulated depreciation account
     */
    accDepreciationAccountId: UUID;
    /**
     * The disposal income account
     */
    disposalIncomeAccountId: UUID;
    /**
     * The disposal expense account
     */
    disposalExpenseAccountId: UUID;
    /**
     * The depreciation expense account
     */
    expenseAccountId: UUID;
};